Add files via upload
This project is to visualize the data taken in a survey of the world by many data analysts, their salaries, job satisfaction, and many other parameters given in the excel sheet attached with this in my repository.This project aims to visualize data collected from a global survey, which includes information on the salaries and job satisfaction of data analysts. The Excel sheet containing the survey data is attached to my repository.