Add files via upload
This is the Excel document used to create the professional survey data breakdown project in Power BI to create a dashboard. This Excel document is utilized for developing the professional survey data breakdown project in Power BI to generate a dashboard. This Excel document is used for creating the professional survey data breakdown project in Power BI to generate a dashboard.